int arg to resize

+++ circlevis/visualizer.py
@@ -58,8 +58,8 @@ class Visualizer(QMainWindow):
         # not meant to be changed to increase the window size (same with width).
         from .renderer import (GAMEPLAY_WIDTH, GAMEPLAY_HEIGHT,
             GAMEPLAY_PADDING_WIDTH, GAMEPLAY_PADDING_HEIGHT)
-        self.resize((GAMEPLAY_WIDTH + GAMEPLAY_PADDING_WIDTH * 2) * 1.2,
-                    (GAMEPLAY_HEIGHT + GAMEPLAY_PADDING_HEIGHT * 2) * 1.2)
+        self.resize((int(GAMEPLAY_WIDTH + GAMEPLAY_PADDING_WIDTH * 2) * 1.2),
+                    int((GAMEPLAY_HEIGHT + GAMEPLAY_PADDING_HEIGHT * 2) * 1.2))
 
     def closeEvent(self, event):
         super().closeEvent(event)
